# Conflicts:
#	src/main/java/com/green/babyfood/admin/AdminController.java
#	src/main/java/com/green/babyfood/admin/AdminService.java

+++ src/main/java/com/green/babyfood/buy/model/BuyPointUpdDto.java
@@ -5,5 +5,5 @@ import lombok.Data;
 @Data
 public class BuyPointUpdDto {
     private Long iuser;
-    private int point;
+    private float point;
 }
